added remarks
