+++ src/main/resources/application.properties
@@ -6,11 +6,6 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
-#spring.liquibase.enabled=true
-#spring.liquibase.change-log=classpath:/db/changelog/db.changelog-master.xml
-
-
-
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.format_sql=true
